+++ src/vm/context.rs
@@ -214,17 +214,16 @@ impl<'runtime, 'program, 'reader, 'parselet> Context<'runtime, 'program, 'reader
         &mut self,
         capture_start: usize, // Stack offset to start from
         copy: bool,           // Copy values instead of draining them from the stack
-        single: bool,         // Lists with single value become a single value, not a list
-        mut inherit: bool,    // Inherit capture for pushing it on the stack again
-        severity: u8,         // Discard any values below given severity
         debug: bool,          // Print debug information
-    ) -> Result<Option<RefValue>, Capture> {
+    ) -> Capture {
+        // Early abort when capture_start is behind stack len
         if capture_start > self.runtime.stack.len() {
-            return Ok(None);
+            return Capture::Empty;
         }
 
         // Eiter copy or drain captures from stack
         let captures: Vec<Capture> = if copy {
+            // fixme: copy feature isn't used...
             Vec::from_iter(
                 self.runtime.stack[capture_start..]
                     .iter()
@@ -240,31 +239,24 @@ impl<'runtime, 'program, 'reader, 'parselet> Context<'runtime, 'program, 'reader
         };
 
         if debug {
-            self.debug(&format!(
-                "collect captures = {} single = {}, severity = {}",
-                captures.len(),
-                single,
-                severity
-            ));
+            self.debug(&format!("collect captures = {}", captures.len(),));
 
             for i in 0..captures.len() {
                 self.debug(&format!(" {}: {:?}", i, captures[i]));
             }
         }
 
-        // No captures, then just stop!
+        // Early abort when no valuable captures had been taken
         if captures.len() == 0 {
-            return Ok(None);
+            return Capture::Empty;
         }
+
+        // Capture inheritance is only possible when there is only one capture available
+        let inherit = captures.len() == 1;
 
         let mut list = List::new();
         let mut dict = Dict::new();
-        let mut max = severity;
-
-        // Capture inheritance is only possible when there is only one capture
-        if inherit && captures.len() > 1 {
-            inherit = false;
-        }
+        let mut max = 0;
 
         // Collect any significant captures and values
         for capture in captures.into_iter() {
@@ -281,8 +273,8 @@ impl<'runtime, 'program, 'reader, 'parselet> Context<'runtime, 'program, 'reader
                     if let Some(alias) = alias {
                         dict.insert(alias, value);
                     } else if inherit {
-                        return Err(Capture::Range(range, alias, severity));
-                    } else {
+                        return Capture::Range(range, alias, severity);
+                    } else if !value.is_void() {
                         list.push(value);
                     }
                 }
@@ -294,14 +286,12 @@ impl<'runtime, 'program, 'reader, 'parselet> Context<'runtime, 'program, 'reader
                         dict.clear();
                     }
 
-                    if !value.is_void() {
-                        if let Some(alias) = alias {
-                            dict.insert(alias, value);
-                        } else if inherit {
-                            return Err(Capture::Value(value, alias, severity));
-                        } else {
-                            list.push(value);
-                        }
+                    if let Some(alias) = alias {
+                        dict.insert(alias, value);
+                    } else if inherit {
+                        return Capture::Value(value, alias, severity);
+                    } else if !value.is_void() {
+                        list.push(value);
                     }
                 }
 
@@ -315,12 +305,12 @@ impl<'runtime, 'program, 'reader, 'parselet> Context<'runtime, 'program, 'reader
         }
 
         if dict.len() == 0 {
-            if list.len() > 1 || (list.len() > 0 && !single) {
-                Ok(Some(RefValue::from(list)))
+            if list.len() > 1 {
+                Capture::Value(RefValue::from(list), None, max)
             } else if list.len() == 1 {
-                Ok(Some(list.pop().unwrap()))
+                Capture::Value(list.pop().unwrap(), None, max)
             } else {
-                Ok(None)
+                Capture::Empty
             }
         } else {
             // Store list-items additionally when there is a dict?
@@ -340,7 +330,7 @@ impl<'runtime, 'program, 'reader, 'parselet> Context<'runtime, 'program, 'reader
                 idx += 1;
             }
 
-            Ok(Some(RefValue::from(dict)))
+            Capture::Value(RefValue::from(dict), None, max)
         }
     }